Platform can change via commands
-Added commands
-Added platformController
-changed double to float

+++ ArmSim/Actuator.cpp
@@ -2,8 +2,8 @@
 
 
 Actuator::Actuator(
-    Platform* _ptrPlatform, int _actuator_ID, ActuatorType _actuatorType,
-    double _posX, double _posY, double _posZ)
+    Platform* _ptrPlatform, uint8_t _actuator_ID, ActuatorType _actuatorType,
+    float _posX, float _posY, float _posZ)
 {
     this->ptrPlatform = _ptrPlatform;
     this->actuator_ID = _actuator_ID;
@@ -19,17 +19,17 @@ Actuator::Actuator(
     this->unit = Units::Unit::Undefined;
 }
 
-double Actuator::getCommandedActuationValue()
+float Actuator::getCommandedActuationValue()
 {
     return this->actuationValue;
 }
 
-void Actuator::setCommandedActuationValue(double newVal)
+void Actuator::setCommandedActuationValue(float newVal)
 {
     this->actuationValue = newVal;
 }
 
-int Actuator::getActuatorID()
+uint8_t Actuator::getActuatorID()
 {
     return this->actuator_ID;
 }
