+++ ArmSim/ActuatorFactory.cpp
@@ -9,9 +9,9 @@ ActuatorFactory::ActuatorFactory(Platform* _platform)
 }
 
 void ActuatorFactory::CreateActuator_AndAddToPlatform(
-	int _actuator_ID,
-	double _posX, double _posY, double _posZ,
-	double _forceVecX, double _forceVecY, double _forceVecZ)
+	uint8_t _actuator_ID,
+	float _posX, float _posY, float _posZ,
+	float _forceVecX, float _forceVecY, float _forceVecZ)
 {
 	Actuator* newActuator = new Actuator_Linear(this->ptrPlatform, _actuator_ID, Actuator::ActuatorType::JAB, _posX, _posY, _posZ, _forceVecX, _forceVecY, _forceVecZ);
