+++ ArmSim/Actuator_Linear.cpp
@@ -2,9 +2,9 @@
 
 
 Actuator_Linear::Actuator_Linear(
-    Platform* _ptrPlatform, int _actuator_ID, ActuatorType _actuatorType,
-    double _posX, double _posY, double _posZ,
-    double _forceVecX, double _forceVecY, double _forceVecZ
+    Platform* _ptrPlatform, uint8_t _actuator_ID, ActuatorType _actuatorType,
+    float _posX, float _posY, float _posZ,
+    float _forceVecX, float _forceVecY, float _forceVecZ
 ) :
     Actuator(_ptrPlatform, _actuator_ID, _actuatorType, _posX, _posY, _posZ),
     forceVecX(_forceVecX), forceVecY(_forceVecY), forceVecZ(_forceVecZ)
@@ -12,7 +12,7 @@ Actuator_Linear::Actuator_Linear(
     
 }
 
-double Actuator_Linear::getRealActuationValue()
+float Actuator_Linear::getRealActuationValue()
 {
     return this->actuationValue;
 }
