+++ ArmSim/PivotPoint.cpp
@@ -1,8 +1,8 @@
 #include "PivotPoint.h"
 
 PivotPoint::PivotPoint(
-    double _posX, double _posY, double _posZ,
-    double _rotAxX, double _rotAxY, double _rotAxZ) :
+    float _posX, float _posY, float _posZ,
+    float _rotAxX, float _rotAxY, float _rotAxZ) :
     posX(_posX), posY(_posY), posZ(_posZ),
     rotAxX(_rotAxX), rotAxY(_rotAxY), rotAxZ(_rotAxZ)
 {}
