+++ ArmSim/Sensor.h
@@ -19,8 +19,8 @@ public:
     Sensor() = delete; // We must call the special constructor
     Sensor(Platform* _ptrPlatform, int _sensor_ID, SensorSource _source);
 
-    virtual double getSensorMeasurement() = 0;
-    virtual double getRealValue() = 0;
+    virtual float getSensorMeasurement() = 0;
+    virtual float getRealValue() = 0;
     
     int getSensorID();
     Units::Unit getUnit();
