+++ ArmSim/Sensor_Actuator.h
@@ -12,8 +12,8 @@ public:
     //Sensor_Actuator() = default; // throws compiler error
     Sensor_Actuator(Platform* _ptrPlatform, int _sensor_ID, Actuator* aPtr);
 
-    double getSensorMeasurement() override;
-    double getRealValue() override;
+    float getSensorMeasurement() override;
+    float getRealValue() override;
 
 
 private:
