+++ ArmSim/Sensor_SystemProperties.h
@@ -13,8 +13,8 @@ class Sensor_SystemProperties : public Sensor
 public:
     Sensor_SystemProperties(Platform* _ptrPlatform, int _sensor_ID, Platform::System_Property _propertyToMeasure, Type_PtrToPlatformMethod funcPtr);
 
-    double getSensorMeasurement() override;
-    double getRealValue() override;
+    float getSensorMeasurement() override;
+    float getRealValue() override;
 
 private:
     Platform::System_Property propertyToMeasure;
